Update web.py
This code works cool, but I have seen a way that we can speed up the process. while brute forcing it the words/second depends on the facebook site that we use. for example, if we use www.facebook.com/login then it will take more time. there are facebook sites for slow connections or basic facebook logins. all of that we are interested in is to find is the password once we find it then we will go a proper site. all in all I have seen that if we change our bruteforcing site to https://touch.facebook.com/login?soft=auth then it will be faster